# services/lab-auth/openapi.yaml
openapi: 3.1.0
info:
  title: Rhine Lab boot identity API
  version: "1.0.0"
  description: >-
    G1 frozen contract for the /lab/ boot identity flow, with the additive
    register-v1 extension (service 0.4.0+, LOGIN-IMPROVE L1a). Implemented by
    services/lab-auth (Go). All paths are exact under /lab/api/auth/. All
    responses are Cache-Control: no-store. Errors never echo passwords or
    distinguish "unknown user" from "wrong password". Registration is not
    login: POST /register creates a persistent account and never sets or
    changes a session cookie.
...
servers:
  - url: /lab/api/auth
paths:
...
  /confirm:
    post:
      summary: Promote a pending attempt to an active session
      operationId: confirm
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfirmRequest" }
      responses:
        "200":
          description: Active session. Never sets a cookie; reuses the pending session cookie.
          headers:
            Cache-Control:
              schema: { type: string, const: "no-store" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfirmResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
components:
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
  schemas:
    User:
      type: object
      required: [id, username]
      properties:
        id: { type: string, description: Stable random user id }
        username:
          {
            type: string,
            description: Canonical stored username (display case),
          }
    Error:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - invalid_credentials
                - origin_rejected
                - csrf_rejected
                - state_conflict
                - payload_too_large
                - rate_limited
                - registration_unavailable
                - registration_disabled
                - unavailable
            message: { type: string }
        requestId: { type: string }
...
    ConfirmRequest:
      type: object
      required: [attemptId]
      properties:
        attemptId: { type: string }
    ConfirmResponse:
      type: object
      required: [authenticated, user, sessionExpiresAt, csrfToken]
      properties:
        authenticated: { type: boolean, const: true }
        user: { $ref: "#/components/schemas/User" }
        sessionExpiresAt: { type: string, format: date-time }
        csrfToken: { type: string, description: Session-bound token for logout }
...
  responses:
    BadRequest:
      description: Malformed body, unknown field, wrong content type, or body over 4 KiB.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Credentials rejected. Never reveals whether the username exists.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Origin or CSRF check failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Attempt already cancelled/confirmed/expired.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    PayloadTooLarge:
      description: Request body exceeds 4 KiB.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Source or username bucket exhausted.
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unavailable:
      description: Verification queue full or storage unavailable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }